[1.12] Fix setting etcd-server when no mTLS is set between etcd and kube-apiserver
An error is introduced in commit 98b0c0a1f1717fd9f926129a681f27eb11d23702 when I cherry pick https://github.com/kubernetes/kubernetes/pull/74690, when mTLS is not enabled, etcd-server was not set, this patch is to fix the issue. This PR is temporarily rolled back in OSS for performance issue investigation, so I cannot fix it in OSS now. I am fixing this in GKE for now, I will add a buganizer to track and make sure when this is fixed in open source, this fix will not carry on.

Change-Id: If9eb6f7d79b9991089e49a7911a07189fa6d57a2