kube-apiserver: Add --ssh-preferred-address-types flag.
(cherry picked from commit 0c9d3455a33006f6e9a15594e85db0fb7430a381)

SSH tunnels are a deprecated feature that is used in GKE to establish
connections from the master to pods running in a different VPC. In this
case, the target of the SSH tunnel is the public IP address of the node.

We also use SSH tunnels in Anthos GKE on AWS Alpha to establish
connections from the master to pods running in a different Calico mesh
(due to Kubeception). In this case, the target of hte SSH tunnel is the
private IP address of the node. But this doesn't work because
kube-apiserver hardcodes logic to look for the "ExternalIP" of the node.

Fix this by adding a new flag --ssh-preferred-address-types that follows
the same convention as --kubelet-preferred-address-types. This new flag
controls which types of addresses the SSH tunnel will attempt to target.

Note: SSH tunnels are deprecated and will not be used in Anthos GKE on
AWS Beta, so we're making a targeted fix here only to for 1.14.7, which
is what Anthos GKE on AWS Alpha ships.

Tested:
* Built new kube-apiserver Docker image.
* Verified that, by default, ExternalIP is still used for SSH tunnels.
* Verified that --ssh-preferred-address-types=InternalIP allows SSH
  tunnels to work in Anthos GKE on AWS, where VMs only have private IPs.

Bug: 146456207
Change-Id: I2767bd3d6962873332f5c938d276c5fa0fb63085
(cherry picked from commit 5281d9b0f108709a2eed489a7fdc4d0637e45486)